QG: fixed warning
git-svn-id: https://coolfluidsrv.vki.ac.be/svn/coolfluid3/trunk@63 4892e656-8fb9-448a-b5bc-b1919152596f

+++ src/Common/tests/TestStringOps.cxx
@@ -74,7 +74,7 @@ BOOST_AUTO_TEST_CASE( boostFunctions )
   
   // getWords
   algorithm::split(myVector, str, algorithm::is_any_of(" "), token_compress_on);
-  BOOST_CHECK_EQUAL(myVector.size(), 5);
+  BOOST_CHECK_EQUAL(myVector.size(), (unsigned int) 5);
   BOOST_CHECK_EQUAL(myVector[0], "");
   BOOST_CHECK_EQUAL(myVector[1], "HeLLo");
   BOOST_CHECK_EQUAL(myVector[2], "World");
